Refactored package structure of HyperBand.
