+++ JAICore/jaicore-ml/src/main/java/ai/libs/jaicore/ml/hpo/multifidelity/hyperband/IHyperBandConfig.java
@@ -1,4 +1,4 @@
-package ai.libs.jaicore.ml.hpo.hyperband;
+package ai.libs.jaicore.ml.hpo.multifidelity.hyperband;
 
 import ai.libs.jaicore.basic.IOwnerBasedAlgorithmConfig;
